Move INIT_STATE on browser part :rowboat:
