+++ mobile/src/reducers/mediaBackup.js
@@ -1,4 +1,4 @@
-import { INIT_STATE } from '../actions'
+import { INIT_STATE } from '../../../src/actions'
 import { MEDIA_UPLOAD_START, MEDIA_UPLOAD_END, MEDIA_UPLOAD_SUCCESS, CURRENT_UPLOAD } from '../actions/mediaBackup'
 
 export const initialState = {
